broker: move RFC 33 validation out of broker
Problem: the broker validates [policy] and [queue] config
tables, but it should not be aware of these constructs.

To preserve separation of concerns, move the code that validates
these tables to libfluxutil and call it from job-manager and job-ingest,
where those config tables are used.  Since job-manager is loaded before
job-ingest, but only on rank 0, checking in both places ensures that any
errors are caught on all ranks, and as early as possible.

Change t2245-policy-config.t from 'minimal' to 'job' personality
so the tests continue to work.  Add a test that ensures a bad config
at instance start is appropriately fatal.
