+++ src/modules/job-ingest/job-ingest.c
@@ -25,6 +25,7 @@
 #include "src/common/libutil/jpath.h"
 #include "src/common/libutil/errprintf.h"
 #include "src/common/libjob/job_hash.h"
+#include "src/common/libfluxutil/policy.h"
 #include "ccan/str/str.h"
 
 #include "util.h"
@@ -647,6 +648,8 @@ static int job_ingest_configure (struct job_ingest_ctx *ctx,
 {
     flux_error_t conf_error;
 
+    if (policy_validate (conf, error) < 0)
+        return -1;
     if (pipeline_configure (ctx->pipeline, conf, argc, argv, error) < 0)
         return -1;
     if (flux_conf_unpack (conf,
